+++ src/fetchez/cli.py
@@ -627,18 +627,22 @@ def fetchez_cli():
             for name, meta in sorted(grouped_hooks[cat], key=lambda x: x[0]):
                 # Grab the cleaned metadata directly from the dictionary
                 desc = meta.get("desc", "No description provided.")
-                mod_path = meta.get("mod", "")
+                # mod_path = meta.get("mod", "")
 
-                # Determine the origin of the hook
-                if mod_path:
-                    origin = mod_path.split(".")[0].capitalize()
-                else:
-                    origin = "User Plugin"
-
+                # # Determine the origin of the hook
+                # if mod_path:
+                #     origin = mod_path.split(".")[0].capitalize()
+                # else:
+                #     origin = "User Plugin"
+                cat_stage = meta.get("stage", "file")
+                if cat_stage == "pre":
+                    cat_stage = "manifest"
+                if cat_stage == "post":
+                    cat_stage = "collection"
                 # Print with origin tag in yellow
                 print(
                     f"  {utils.colorize(name, utils.BOLD):<26} "
-                    f"{utils.colorize(f'[{origin}]', utils.YELLOW):<13} : {desc}"
+                    f"{utils.colorize(f'[{cat_stage}]', utils.YELLOW):<13} : {desc}"
                 )
 
         print()