+++ CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+
+### Changed
+
+### BUGFIX
+
+## [0.5.2] - 2026-05-02
 ### ADDED
 - Add "lidarbc" fetchez modules (canada)
 - Add .fetchez_cache to save module results for re-use